requirements.txt
Django~=4.0.2
whitenoise~=6.4.0
djangorestframework~=3.14.0
python-decouple~=3.8
virustotal_python~=1.0.2
Pillow~=8.1.1
tensorflow~=2.11.0
numpy~=1.24.1